@@ -1,16 +1,9 @@
+/* eslint-disable import/no-unresolved */
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
+import { createRoot } from 'react-dom/client';
 
 import App from './App';
-import ThemeContextProvider from './shared/context/theme';
-import store from './store/index.store';
 
-ReactDOM.render(
-  <ThemeContextProvider>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </ThemeContextProvider>,
-  document.querySelector('#root'),
-);
+const container = document.getElementById('root') as HTMLElement;
+const root = createRoot(container);
+root.render(<App />);
