@@ -1,11 +1,12 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
 export const FormContainer = styled.form`
-  min-width: 300px;
   display: flex;
   gap: 0.8rem;
-  flex-direction: column;
   background-color: ${props => props.theme.secondary};
+  justify-content: flex-start;
+  align-content: space-around;
+  flex-wrap: wrap;
   padding: 1rem;
   box-shadow: 0.5px 0.6px 0.9px hsl(0deg 0% 63% / 0.34),
     0.8px 0.9px 1.4px -1.2px hsl(0deg 0% 63% / 0.34),
@@ -13,8 +14,13 @@ export const FormContainer = styled.form`
 `;
 
 export const FormButton = styled.button`
-  padding: 0.5rem;
+  ${({ theme }) => css`
+    background-color: ${theme.colors.blue};
+    color: ${theme.text.alt};
+  `}
+
+  padding: 10px;
+  font-weight: 600;
   border: none;
   border-radius: 0.2rem;
-  background-color: ${props => props.theme.colors.blue};
 `;
