@@ -2,8 +2,6 @@ import styled from 'styled-components';
 
 export const WalletMain = styled.section`
   display: flex;
-  align-items: flex-start;
-  padding-right: 0.3rem;
-  margin-top: 2rem;
-  gap: 1rem;
+  flex-direction: column;
+  padding: 10px;
 `;
