@@ -1,5 +1,5 @@
 import React from 'react';
-import { FieldErrors, UseFormRegister } from 'react-hook-form';
+import { UseFormRegister } from 'react-hook-form';
 
 import { TFormData } from '../../../core/Form';
 import * as S from './style';
@@ -9,22 +9,20 @@ type SelectProps = {
   options: string[];
   name: 'value' | 'description' | 'currency' | 'method' | 'tag';
   register: UseFormRegister<TFormData>;
-  errors: FieldErrors<TFormData>;
 };
 
-function Select({ label, options, name, register, errors }: SelectProps) {
+function Select({ label, options, name, register }: SelectProps) {
   return (
     <S.SelectControl>
       <S.Label htmlFor={name}>{label}</S.Label>
       <S.Select
-        {...register(name, {
-          required: true,
-        })}
+        {...register(name)}
         required
         name={name}
         id={name}
+        defaultValue={options[0]}
       >
-        <option hidden disabled></option>
+        <option hidden disabled />
         {options.map(each => (
           <option key={each} value={each}>
             {each}
