@@ -3,6 +3,7 @@ import styled from 'styled-components';
 export const SelectControl = styled.div`
   display: flex;
   flex-direction: column;
+  gap: 5px;
   align-items: flex-start;
 `;
 
@@ -13,8 +14,7 @@ export const Label = styled.label`
 
 export const Select = styled.select`
   flex-grow: 0;
-  padding: 0.1rem;
-  width: 100px;
+  padding: 0.5rem;
   color: ${({ theme }) => theme.text.default};
   background-color: ${({ theme }) => theme.secondary};
   filter: brightness(0.8);
