@@ -1,3 +1,3 @@
-export * from './Card';
 export * from './Input';
 export * from './Select';
+export * from './TableTools';
