@@ -1,8 +1,5 @@
-/* eslint-disable unicorn/no-unsafe-regex */
 export const validations = (email: string, password: string) => {
-  const regex = /^[\w-\\.]+@([\w-]+.)+[\w-]{2,4}$/g;
+  const regex = /^[\w-\\.]+@([\w-]+.)+[\w-]{2,4}$/g; // eslint-disable-line security/detect-unsafe-regex
   const MIN_CARACTERES = 5;
   return !(regex.test(email) && password.length > MIN_CARACTERES);
 };
-
-export const nada = 'nada';
