@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ThemeContext } from '../../context/theme';
+import { ThemeContext } from '../context/theme';
 
 const useTheme = () => {
   return useContext(ThemeContext);
