@@ -1,6 +1,6 @@
-import {createSlice, PayloadAction} from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-type TInitialState = {email: string};
+type TInitialState = { email: string };
 
 const initialState: TInitialState = {
   email: '',
@@ -10,12 +10,12 @@ const user = createSlice({
   name: 'user',
   initialState,
   reducers: {
-    actionSetEmail(state, {payload}: PayloadAction<string>) {
-      state.email = payload;
+    actionSetEmail(state, { payload }: PayloadAction<string>) {
+      state.email = payload; // eslint-disable-line no-param-reassign
     },
   },
 });
 
-export const {actionSetEmail} = user.actions;
+export const { actionSetEmail } = user.actions;
 
 export default user.reducer;
